🎫 Limit to admin patch/delete user

+++ src/routes/_api/users/[_id]/+server.js
@@ -1,10 +1,18 @@
 import { userService } from "$lib/server/services/user.service";
-import { json } from "@sveltejs/kit";
+import { error, json } from "@sveltejs/kit";
 
 /**
  * @type {import("./$types").RequestHandler}
  */
 export const PATCH = async ({ params, request }) => {
+	const token = request.headers.get("authorization").split(" ")[1];
+	const user = await userService.getFromJWT(token);
+
+	// @ts-ignore
+	if (!user || !user.profiles.includes("admin")) {
+		throw error(403, { id: "user.unauthorized", message: "Unauthorized" });
+	}
+
 	const body = await request.json();
 	const result = await userService.update(params._id, body);
 
@@ -14,7 +22,15 @@ export const PATCH = async ({ params, request }) => {
 /**
  * @type {import("./$types").RequestHandler}
  */
-export const DELETE = async ({ params }) => {
+export const DELETE = async ({ params, request }) => {
+	const token = request.headers.get("authorization").split(" ")[1];
+	const user = await userService.getFromJWT(token);
+
+	// @ts-ignore
+	if (!user || !user.profiles.includes("admin")) {
+		throw error(403, { id: "user.unauthorized", message: "Unauthorized" });
+	}
+
 	const result = await userService.delete(params._id);
 
 	return json(result);
